MCR-3650 refactor MCRConfigurableInstanceHelper into multiple classes
