+++ mycore-base/src/main/java/org/mycore/common/config/MCRConfiguration2.java
@@ -32,6 +32,8 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import org.mycore.common.MCRClassTools;
+import org.mycore.common.config.MCRInstanceConfiguration.Option;
+import org.mycore.common.config.MCRInstanceConfiguration.Options;
 import org.mycore.common.function.MCRTriConsumer;
 
 /**
@@ -131,11 +133,11 @@ public class MCRConfiguration2 {
      * @throws MCRConfigurationException if the class can not be loaded or instantiated
      */
     public static <S> Optional<S> getInstanceOf(Class<S> superClass, String name) throws MCRConfigurationException {
-        return getInstanceOf(superClass, name, MCRConfigurableInstanceHelper.NO_OPTIONS);
+        return getInstanceOf(superClass, name, Options.NONE);
     }
 
     private static <S> Optional<S> getInstanceOf(Class<S> superClass, String name,
-        Set<MCRConfigurableInstanceHelper.Option> options) {
+        Set<Option> options) {
         if (MCRConfigurableInstanceHelper.isSingleton(superClass)) {
             return getSingleInstanceOf(superClass, name, options);
         } else {
@@ -153,7 +155,7 @@ public class MCRConfiguration2 {
      * or the configuration property is not set
      */
     public static <S> S getInstanceOfOrThrow(Class<S> superClass, String name) throws MCRConfigurationException {
-        return getInstanceOf(superClass, name, MCRConfigurableInstanceHelper.ADD_IMPLICIT_CLASS_PROPERTIES)
+        return getInstanceOf(superClass, name, Options.IMPLICIT)
             .orElseThrow(() -> createConfigurationException(name));
     }
 
@@ -167,11 +169,11 @@ public class MCRConfiguration2 {
      * @throws MCRConfigurationException if the class can not be loaded or instantiated
      */
     public static <S> Optional<S> getSingleInstanceOf(Class<S> superClass, String name) {
-        return getSingleInstanceOf(superClass, name, MCRConfigurableInstanceHelper.NO_OPTIONS);
+        return getSingleInstanceOf(superClass, name, Options.NONE);
     }
 
     private static <S> Optional<S> getSingleInstanceOf(Class<S> superClass, String name,
-        Set<MCRConfigurableInstanceHelper.Option> options) {
+        Set<Option> options) {
         return getString(name)
             .map(className -> new ConfigSingletonKey(name, className))
             .map(key -> (S) instanceHolder.computeIfAbsent(key,
@@ -189,7 +191,7 @@ public class MCRConfiguration2 {
      * or the configuration property is not set
      */
     public static <S> S getSingleInstanceOfOrThrow(Class<S> superClass, String name) {
-        return getSingleInstanceOf(superClass, name, MCRConfigurableInstanceHelper.ADD_IMPLICIT_CLASS_PROPERTIES)
+        return getSingleInstanceOf(superClass, name, Options.IMPLICIT)
             .orElseThrow(() -> createConfigurationException(name));
     }
