+++ mycore-base/src/main/java/org/mycore/common/config/annotation/MCRProperty.java
@@ -67,7 +67,7 @@ public @interface MCRProperty {
      * @return The {@link MCRSentinel} for the configured property.
      */
     MCRSentinel sentinel() default @MCRSentinel(enabled = false);
-    
+
     /**
      * @return The order in which the annotated fields or methods are processed. The higher the value, the later the
      * field or method is processed. All fields are processed first, then all methods are processed.
