+++ mycore-base/src/test/java/org/mycore/common/config/MCRConfigurableInstanceHelperBasicTest.java
@@ -26,6 +26,7 @@ import java.util.function.Supplier;
 import org.junit.jupiter.api.Test;
 import org.mycore.common.MCRTestConfiguration;
 import org.mycore.common.MCRTestProperty;
+import org.mycore.common.config.MCRInstanceConfiguration.Options;
 import org.mycore.common.config.annotation.MCRConfigurationProxy;
 import org.mycore.common.config.annotation.MCRFactory;
 import org.mycore.test.MyCoReTest;
@@ -135,7 +136,7 @@ public class MCRConfigurableInstanceHelperBasicTest {
 
         MCRInstanceConfiguration configuration = MCRInstanceConfiguration.ofName("Foo");
         ImplicitTestClass instance = MCRConfigurableInstanceHelper.getInstance(ImplicitTestClass.class, configuration,
-            MCRConfigurableInstanceHelper.ADD_IMPLICIT_CLASS_PROPERTIES);
+            Options.IMPLICIT);
 
         assertNotNull(instance);
